refactoring of seeders
